TA93957 - US107407 - Identity Resolution Guard
Breaks the identity resolution functionality of the AIMSRouteGuard into
its own guard service.  This allows `aaid` and `locid` parameters to be
converted into the current execution context before a given route is
activated.

Also adds query parameter aggregation in alNavigationService (for later
use) and fixes an annoying defect in
`AlNavigationService.setActingAccount()` that caused a redirection error
while running on localhost.